wip(refactoring): smartlink, availability, renderField, handleSmartlinkClick extracted to separate files and scopes
