+++ src/components/SmartLink/index.ts
@@ -1,0 +1,1 @@
+export { SmartLink } from "./SmartLink";
